Some test functionality organization.
